Try calling for search results in inline script
