+++ web/public/index.html
@@ -59,6 +59,14 @@
       }
     </style>
 
+    <% if (process.env.REACT_APP_WHICH_APP === "search"){ %>
+    <!--
+      @gladly-customization:
+      YPA script for search results.
+    -->
+    <script src="https://s.yimg.com/uv/dm/scripts/syndication.js" />
+    <% } %>
+
     <!--
       @gladly-customization:
       The new tab page may be displayed in an iframe, so open
